bi_tool/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
djongo==1.3.6
pymongo==4.6.0
python-decouple==3.8
django-cors-headers==4.3.1
dnspython==2.4.2
sqlparse==0.4.4
pytz==2023.3

# Additional packages for analytics and business logic
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
python-dateutil==2.8.2

# ETL and Data Processing
django-celery-beat==2.5.0
flower==2.0.1
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
boto3==1.34.0
sentry-sdk==1.38.0
prometheus-client==0.19.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0

# Data Validation and Processing
pydantic==2.5.0
jsonschema==4.20.0
xlrd==2.0.1
openpyxl==3.1.2
python-magic==0.4.27
chardet==5.2.0

# Task Queue and Background Processing
django-rq==2.8.1
rq-scheduler==0.13.1

# API and Documentation
drf-yasg==1.21.7
swagger-spec-validator==3.0.3